completed sales report 2025-04-09

+++ backend/database/migrations/2025_03_09_131857_create_stocks_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('batch_number');
             $table->integer('quantity');
             $table->integer('unit_price');
-            $table->integer('total_cost');
+            $table->bigInteger('total_cost');
             $table->integer('sale_price');
             $table->integer('min_stock_level');
             $table->foreignId('supplier_id')->constrained('suppliers')->cascadeOnDelete();
